requirements.txt
torch==2.1.0+cu121
torchvision==0.16.0+cu121

# albumentations==1.4.4
# einops==0.7.0
matplotlib==3.7.1
numpy==1.24.3
opencv_python==4.8.1.78
pandas==1.5.3
Pillow==9.4.0
psutil==5.9.0
# pycocotools==2.0.7
PyYAML==6.0
scipy==1.10.1
seaborn==0.13.2
setuptools>=68.0.0
thop==0.1.1.post2209072238
timm==0.9.16
tqdm==4.65.0

onnxruntime==1.16.3
# pycuda==2024.1
# tensorrt==10.0.1
ppq==0.6.6

ultralytics==8.2.2
